Add --remove option

+++ imagepile/validate.py
@@ -4,13 +4,14 @@ import numpy as np
 from PIL import Image
 from concurrent.futures import ProcessPoolExecutor
 from pathlib import Path
+from functools import partial
 
 logger = logging.getLogger(__name__)
 
 EXTENSIONS=['.jpg', '.jpeg', '.png']
 
 
-def _validate_image(filepath):
+def _validate_image(filepath, remove=False):
     try:
         with Image.open(filepath) as img:
             img.convert("RGB").verify()
@@ -18,22 +19,24 @@ def _validate_image(filepath):
         return True
     except Exception as e:
         print(f"[{e.__class__.__name__}] {e}: {str(filepath)}")
-        os.remove(filepath)
+        if remove:
+            os.remove(filepath)
         return False
 
 
-def validate(folderpath):
+def validate(folderpath, remove=False):
     root = Path(folderpath)
     n = len(root.parts)
     filepaths = np.array([
         str(Path(*f.parts[n:])) for f in root.rglob("*")
         if f.is_file() and f.suffix in EXTENSIONS
     ])
-    logger.info(f"Num filepaths {len(filepaths)}")
+    logger.info(f"Number of images {len(filepaths)}")
 
     with ProcessPoolExecutor(max_workers=os.cpu_count()) as executor:
         mask = np.array(list(executor.map(
-            _validate_image, (root / Path(f) for f in filepaths))))
+            partial(_validate_image, remove=remove),
+            (root / Path(f) for f in filepaths))))
 
     filepaths = list(filepaths[mask])
-    logger.info(f"Num filepaths {len(filepaths)}")
+    logger.info(f"Number of filtered images: {len(filepaths)}")
